+++ src/REvoDesign/citations/__init__.py
@@ -1,6 +1,6 @@
 '''
 Module to import all citation modules
 '''
-from .CitationManager import CitableModules, CitationManager
+from .CitationManager import CitableModuleAbstract, CitationManager
 
-__all__ = ["CitationManager", "CitableModules"]
+__all__ = ["CitationManager", "CitableModuleAbstract"]