+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - test case, which were mostly written by prompting with ChatGPT and Tongyi Lingma:
   - basic tests
   - bootstrap tests
@@ -29,6 +30,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - utils: `cmap_reverser`, `count_and_sort_characters`, `generate_strong_password`, `get_color`, `random_deduplicate`, `rescale_number`, `timing`,`extract_archive`, `minibatches`, `minibatches_generator`
 
 ### Changed
+
+- renamed:
+  - `CitableModules` -> `CitableModuleAbstract`: Explicitly named as an abstract class.
 - refactored `SingletonAbstract`:
   - `singleton_init`: an abstract method for each subclass to implement. This method will be called when the subclass is initialized, helps to avoid repeated code duplication. This method is the **only** one developer should implement in subclasses to replace the original `__init__` method.
   - `derive`: derive a new singleton class from the current one.
@@ -41,7 +45,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ### Removed
+
 - installer: install from local git: currently not working.
+- Makefile:
+  - removed pytest args: `-x` to continue on error
 
 ## [1.7.8] - 2024-12-11